refactor: rename constants
