+++ src/logic/contracts/methodIds.ts
@@ -1,15 +1,20 @@
 import { web3ReadOnly as web3 } from 'src/logic/wallets/getWeb3'
-import { DataDecoded, METHOD_TO_ID } from 'src/logic/safe/store/models/types/transactions.d'
+import {
+  DataDecoded,
+  SAFE_METHOD_ID_TO_NAME,
+  SPENDING_LIMIT_METHOD_ID_TO_NAME,
+  SPENDING_LIMIT_METHODS_NAMES,
+} from 'src/logic/safe/store/models/types/transactions.d'
 
 export const decodeParamsFromSafeMethod = (data: string): DataDecoded | null => {
-  const [methodId, params] = [data.slice(0, 10) as keyof typeof METHOD_TO_ID | string, data.slice(10)]
+  const [methodId, params] = [data.slice(0, 10) as keyof typeof SAFE_METHOD_ID_TO_NAME | string, data.slice(10)]
 
   switch (methodId) {
     // swapOwner
     case '0xe318b52b': {
       const decodedParameters = web3.eth.abi.decodeParameters(['uint', 'address', 'address'], params) as string[]
       return {
-        method: METHOD_TO_ID[methodId],
+        method: SAFE_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: 'oldOwner', type: 'address', value: decodedParameters[1] },
           { name: 'newOwner', type: 'address', value: decodedParameters[2] },
@@ -21,7 +26,7 @@ export const decodeParamsFromSafeMethod = (data: string): DataDecoded | null => 
     case '0x0d582f13': {
       const decodedParameters = web3.eth.abi.decodeParameters(['address', 'uint'], params)
       return {
-        method: METHOD_TO_ID[methodId],
+        method: SAFE_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: 'owner', type: 'address', value: decodedParameters[0] },
           { name: '_threshold', type: 'uint', value: decodedParameters[1] },
@@ -33,7 +38,7 @@ export const decodeParamsFromSafeMethod = (data: string): DataDecoded | null => 
     case '0xf8dc5dd9': {
       const decodedParameters = web3.eth.abi.decodeParameters(['address', 'address', 'uint'], params)
       return {
-        method: METHOD_TO_ID[methodId],
+        method: SAFE_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: 'oldOwner', type: 'address', value: decodedParameters[1] },
           { name: '_threshold', type: 'uint', value: decodedParameters[2] },
@@ -45,7 +50,7 @@ export const decodeParamsFromSafeMethod = (data: string): DataDecoded | null => 
     case '0x694e80c3': {
       const decodedParameters = web3.eth.abi.decodeParameters(['uint'], params)
       return {
-        method: METHOD_TO_ID[methodId],
+        method: SAFE_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: '_threshold', type: 'uint', value: decodedParameters[0] },
         ],
@@ -56,7 +61,7 @@ export const decodeParamsFromSafeMethod = (data: string): DataDecoded | null => 
     case '0x610b5925': {
       const decodedParameters = web3.eth.abi.decodeParameters(['address'], params)
       return {
-        method: METHOD_TO_ID[methodId],
+        method: SAFE_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: 'module', type: 'address', value: decodedParameters[0] },
         ],
@@ -67,7 +72,7 @@ export const decodeParamsFromSafeMethod = (data: string): DataDecoded | null => 
     case '0xe009cfde': {
       const decodedParameters = web3.eth.abi.decodeParameters(['address', 'address'], params)
       return {
-        method: METHOD_TO_ID[methodId],
+        method: SAFE_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: 'prevModule', type: 'address', value: decodedParameters[0] },
           { name: 'module', type: 'address', value: decodedParameters[1] },
@@ -80,32 +85,20 @@ export const decodeParamsFromSafeMethod = (data: string): DataDecoded | null => 
   }
 }
 
-export const SPENDING_LIMIT_METHODS_NAMES = {
-  ADD_DELEGATE: 'addDelegate',
-  SET_ALLOWANCE: 'setAllowance',
-  EXECUTE_ALLOWANCE_TRANSFER: 'executeAllowanceTransfer',
-}
-
-export const SPENDING_LIMIT_METHOD_TO_ID = {
-  '0xe71bdf41': SPENDING_LIMIT_METHODS_NAMES.ADD_DELEGATE,
-  '0xbeaeb388': SPENDING_LIMIT_METHODS_NAMES.SET_ALLOWANCE,
-  '0x4515641a': SPENDING_LIMIT_METHODS_NAMES.EXECUTE_ALLOWANCE_TRANSFER,
-}
-
 export const isSetAllowanceMethod = (data: string): boolean => {
-  const methodId = data.slice(0, 10) as keyof typeof SPENDING_LIMIT_METHOD_TO_ID
-  return SPENDING_LIMIT_METHOD_TO_ID[methodId] === SPENDING_LIMIT_METHODS_NAMES.SET_ALLOWANCE
+  const methodId = data.slice(0, 10) as keyof typeof SPENDING_LIMIT_METHOD_ID_TO_NAME
+  return SPENDING_LIMIT_METHOD_ID_TO_NAME[methodId] === SPENDING_LIMIT_METHODS_NAMES.SET_ALLOWANCE
 }
 
 export const decodeParamsFromSpendingLimit = (data: string): DataDecoded | null => {
-  const [methodId, params] = [data.slice(0, 10) as keyof typeof SPENDING_LIMIT_METHOD_TO_ID | string, data.slice(10)]
+  const [methodId, params] = [data.slice(0, 10) as keyof typeof SPENDING_LIMIT_METHOD_ID_TO_NAME | string, data.slice(10)]
 
   switch (methodId) {
     // addDelegate
     case '0xe71bdf41': {
       const decodedParameter = (web3.eth.abi.decodeParameter('address', params) as unknown) as string
       return {
-        method: SPENDING_LIMIT_METHOD_TO_ID[methodId],
+        method: SPENDING_LIMIT_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: 'delegate', type: 'address', value: decodedParameter },
         ],
@@ -116,7 +109,7 @@ export const decodeParamsFromSpendingLimit = (data: string): DataDecoded | null 
     case '0xbeaeb388': {
       const decodedParameters = web3.eth.abi.decodeParameters(['address', 'address', 'uint96', 'uint16', 'uint32'], params)
       return {
-        method: SPENDING_LIMIT_METHOD_TO_ID[methodId],
+        method: SPENDING_LIMIT_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: 'delegate', type: 'address', value: decodedParameters[0] },
           { name: 'token', type: 'address', value: decodedParameters[1] },
@@ -131,7 +124,7 @@ export const decodeParamsFromSpendingLimit = (data: string): DataDecoded | null 
     case '0x4515641a': {
       const decodedParameters = web3.eth.abi.decodeParameters(['address', 'address', 'address', 'uint96', 'address', 'uint96', 'address', 'bytes'], params)
       return {
-        method: SPENDING_LIMIT_METHOD_TO_ID[methodId],
+        method: SPENDING_LIMIT_METHOD_ID_TO_NAME[methodId],
         parameters: [
           { name: "safe", type: "address", value: decodedParameters[0] },
           { name: "token", type: "address", value: decodedParameters[1] },
@@ -151,11 +144,11 @@ export const decodeParamsFromSpendingLimit = (data: string): DataDecoded | null 
 }
 
 const isSafeMethod = (methodId: string): boolean => {
-  return !!METHOD_TO_ID[methodId]
+  return !!SAFE_METHOD_ID_TO_NAME[methodId]
 }
 
 const isSpendingLimitMethod = (methodId: string): boolean => {
-  return !!SPENDING_LIMIT_METHOD_TO_ID[methodId]
+  return !!SPENDING_LIMIT_METHOD_ID_TO_NAME[methodId]
 }
 
 export const decodeMethods = (data: string): DataDecoded | null => {
